requirements.txt
bidict==0.21.2
click==8.0.0
dataclasses==0.8
Flask==2.0.0
Flask-Cors==3.0.10
Flask-Login==0.5.0
Flask-SocketIO==5.0.1
future==0.18.2
itsdangerous==2.0.0
Jinja2==3.0.0
lxml==4.6.3
MarkupSafe==2.0.0
numpy==1.19.5
opencv-python==4.5.2.52
pkg-resources==0.0.0
pymavlink==2.4.15
pymongo==3.11.4
python-dotenv==0.17.1
python-engineio==4.1.0
python-socketio==5.2.1
six==1.16.0
Werkzeug==2.0.0
